requirements.txt
Django==4.0.2
djangorestframework==3.13.1
psycopg2-binary==2.9.1
isort==4.3.21
gunicorn==20.0.4
django-stubs==1.8.0
django-stubs-ext==0.2.0
django_phonenumber_field==6.1.0
phonenumbers==8.12.44